Move 11ty stub function to a helper file
